+++ sentry-ruby/lib/sentry/net/http.rb
@@ -31,7 +31,10 @@ module Sentry
 
         Sentry.with_child_span(op: OP_NAME, start_timestamp: Sentry.utc_now.to_f) do |sentry_span|
           request_info = extract_request_info(req)
-          set_sentry_trace_header(req, sentry_span, request_info)
+
+          if propagate_trace?(request_info[:url], Sentry.configuration)
+            set_propagation_headers(req)
+          end
 
           super.tap do |res|
             record_sentry_breadcrumb(request_info, res)
@@ -49,17 +52,8 @@ module Sentry
 
       private
 
-      def set_sentry_trace_header(req, sentry_span, request_info)
-        return unless sentry_span
-
-        client = Sentry.get_current_client
-        return unless propagate_trace?(request_info[:url], client.configuration.trace_propagation_targets)
-
-        trace = client.generate_sentry_trace(sentry_span)
-        req[SENTRY_TRACE_HEADER_NAME] = trace if trace
-
-        baggage = client.generate_baggage(sentry_span)
-        req[BAGGAGE_HEADER_NAME] = baggage if baggage && !baggage.empty?
+      def set_propagation_headers(req)
+        Sentry.get_trace_propagation_headers&.each { |k, v| req[k] = v }
       end
 
       def record_sentry_breadcrumb(request_info, res)
@@ -96,8 +90,10 @@ module Sentry
         result
       end
 
-      def propagate_trace?(url, trace_propagation_targets)
-        url && trace_propagation_targets.any? { |target| url.match?(target) }
+      def propagate_trace?(url, configuration)
+        url &&
+          configuration.propagate_traces &&
+          configuration.trace_propagation_targets.any? { |target| url.match?(target) }
       end
     end
   end